refactored modules 2&4 in uniform manner
