--- src/components/Training/module-2/components/index.js
@@ -1,6 +1,0 @@
-import { SimpleStep } from './SimpleStep'
-import { Module21 } from './module2-1'
-import { Module22 } from './module2-2'
-import { Module23 } from './module2-3'
-
-export { SimpleStep, Module21, Module22, Module23 }
