--- src/components/Training/module-4/components/index.js
@@ -1,6 +1,0 @@
-import { Module41 } from './module4-1'
-import { Module42 } from './module4-2'
-import { Module43 } from './module4-3'
-import { Module44 } from './module4-4'
-
-export { Module41, Module42, Module43, Module44 }
