+++ src/components/Training/module-4/components/module4-1.js
@@ -20,8 +20,33 @@ import {
   Header2,
 } from '../../TrainStyles.js'
 import Player from 'react-player'
-import Diamond from './diamond'
+// import Diamond from './diamond'
 import Progress from "../../TrainingModuleContainer";
+
+export const Diamond = ({
+  style = {},
+  fill = '#000',
+  width = '100%',
+  className = '',
+  viewBox = '0 0 32 32',
+}) => (
+  <svg
+    width='9'
+    height='9'
+    viewBox='0 0 12 12'
+    fill='none'
+    xmlns='http://www.w3.org/2000/svg'
+  >
+    <rect
+      x='6'
+      y='0.343262'
+      width='7'
+      height='7'
+      transform='rotate(45 6 0.343262)'
+      fill='#343434'
+    />
+  </svg>
+)
 
 export const Module41 = ({ handleNext, handleBack }) => {
 
@@ -218,3 +243,5 @@ export const Module41 = ({ handleNext, handleBack }) => {
     </>
   )
 }
+
+export default Module41;
