+++ src/components/Training/module-4/components/module4-2.js
@@ -207,3 +207,5 @@ export const Module42 = ({ handleNext, handleBack }) => {
     </>
   )
 }
+
+export default Module42;
