+++ src/components/Training/module-4/components/module4-3.js
@@ -153,3 +153,5 @@ export const Module43 = ({ handleBack, handleNext }) => {
     </>
   )
 }
+
+export default Module43;
