+++ src/components/Training/module-4/components/module4-4.js
@@ -146,3 +146,5 @@ export const Module44 = ({ handleBack }) => {
     </>
   )
 }
+
+export default Module44;
